// admin/src/client/types.gen.ts
export type Assignment = {
    /**
     * Id
     */
    id: string;
    /**
     * Type
     */
    type: string;
    /**
     * Status
     */
    status?: string;
    /**
     * Query Text
     */
    query_text: string;
    /**
     * Query Params
     */
    query_params?: {
        [key: string]: unknown;
    } | null;
    /**
     * Triggered By
     */
    triggered_by?: string | null;
    /**
     * Triggered By Id
     */
    triggered_by_id?: string | null;
    /**
     * Triggered At
     */
    triggered_at?: string;
    /**
     * Parent Assignment
     */
    parent_assignment?: string | null;
    /**
     * Artifacts In
     */
    artifacts_in?: Array<string> | null;
    /**
     * Artifacts Out
     */
    artifacts_out?: Array<string> | null;
    /**
     * Sub Assignments
     */
    sub_assignments?: Array<string> | null;
    /**
     * Monitor
     */
    monitor?: {
        [key: string]: unknown;
    } | null;
};